Remove alternative-increment music property
Instead of computing the length of the volta-numbers property and
sending it in the alternative-increment property, just send the
volta-numbers property and let the receiver compute the length.

+++ lily/include/timing-translator.hh
@@ -44,8 +44,8 @@ private:
   // alt_... members pertain to bar numbering for repeat alternatives
   Stream_event *alt_event_ = nullptr;
   int alt_starting_bar_number_ = 0;
-  int alt_number_ = 0;
-  int alt_number_increment_ = 0;
+  long alt_number_ = 0;
+  long alt_number_increment_ = 0;
   bool alt_reset_enabled_ = false;
 };
